moves hes_datasets to raw_data.mitigators.ip, because that is the only place this is used
