+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/ambulatory_care_conditions.py
@@ -61,7 +61,8 @@ from functools import reduce
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import (
+from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import (
     any_diagnosis,
     any_procedure,
     nhp_apc,
@@ -69,7 +70,6 @@ from nhp.data.hes_datasets import (
     primary_procedure,
     secondary_diagnosis,
 )
-from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
 
 
 @activity_avoidance_mitigator()
