+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/cancelled_operations.py
@@ -24,8 +24,8 @@ HRG codes are published annually by [NHS England][co_1].
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import nhp_apc
 from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import nhp_apc
 
 
 @activity_avoidance_mitigator()
