+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/eol_care.py
@@ -17,8 +17,12 @@ indication that the patient experienced any trauma and the patient dies in hospi
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import any_diagnosis, nhp_apc, procedures
 from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import (
+    any_diagnosis,
+    nhp_apc,
+    procedures,
+)
 
 
 # base function for the two mitigators
