+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/evidence_based_interventions.py
@@ -6,7 +6,7 @@ from typing import Callable
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import any_diagnosis
+from nhp.data.raw_data.mitigators.ip.hes_datasets import any_diagnosis
 
 
 def evidence_based_interventions(*args: Callable[[], DataFrame]) -> DataFrame:
