+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/frail_elderly.py
@@ -38,8 +38,8 @@ import pyspark.sql.types as T
 from pyspark.sql import functions as F
 
 from nhp.data.get_spark import get_spark
-from nhp.data.hes_datasets import diagnoses, nhp_apc
 from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import diagnoses, nhp_apc
 from nhp.data.table_names import table_names
