+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/medicines_related_admissions.py
@@ -36,13 +36,13 @@ set.
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import (
+from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import (
     any_diagnosis,
     nhp_apc,
     primary_diagnosis,
     secondary_diagnosis,
 )
-from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
 
 
 def _medicines_related_admissions_implicit(primary_diags, secondary_diags):
