+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/raid_ae.py
@@ -16,8 +16,12 @@ not undergo a procedure.
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import nhp_apc, primary_diagnosis, procedures
 from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import (
+    nhp_apc,
+    primary_diagnosis,
+    procedures,
+)
 
 
 @activity_avoidance_mitigator()
