+++ src/nhp/data/raw_data/mitigators/ip/activity_avoidance/readmission_within_28_days.py
@@ -13,8 +13,8 @@ readmitted to.
 from pyspark.sql import functions as F
 
 from nhp.data.get_spark import get_spark
-from nhp.data.hes_datasets import nhp_apc
 from nhp.data.raw_data.mitigators import activity_avoidance_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import nhp_apc
 from nhp.data.table_names import table_names
