+++ src/nhp/data/raw_data/mitigators/ip/efficiency/emergency_elderly.py
@@ -10,8 +10,8 @@ In addition minimising the LOS for such patients would also free up capacity wit
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import nhp_apc
 from nhp.data.raw_data.mitigators import efficiency_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import nhp_apc
 
 
 @efficiency_mitigator()
