+++ src/nhp/data/raw_data/mitigators/ip/efficiency/raid_ip.py
@@ -13,8 +13,8 @@ diagnosis.
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import any_diagnosis, nhp_apc
 from nhp.data.raw_data.mitigators import efficiency_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import any_diagnosis, nhp_apc
 
 
 @efficiency_mitigator()
