+++ src/nhp/data/raw_data/mitigators/ip/efficiency/same_day_emergency_care.py
@@ -31,8 +31,8 @@ import pandas as pd
 from pyspark.sql import functions as F
 
 from nhp.data.get_spark import get_spark
-from nhp.data.hes_datasets import diagnoses, nhp_apc
 from nhp.data.raw_data.mitigators import efficiency_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import diagnoses, nhp_apc
 from nhp.data.raw_data.mitigators.reference_data import load_json
 from nhp.data.table_names import table_names
