+++ src/nhp/data/raw_data/mitigators/ip/efficiency/stroke_early_supported_discharge.py
@@ -12,8 +12,8 @@ inpatient LOS through supported discharge.
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import nhp_apc
 from nhp.data.raw_data.mitigators import efficiency_mitigator
+from nhp.data.raw_data.mitigators.ip.hes_datasets import nhp_apc
 
 
 @efficiency_mitigator()
