+++ src/nhp/data/raw_data/mitigators/ip/shared/virtual_wards.py
@@ -13,7 +13,11 @@ Infections (ARI) and Heart failure.
 
 from pyspark.sql import functions as F
 
-from nhp.data.hes_datasets import nhp_apc, primary_diagnosis, procedures
+from nhp.data.raw_data.mitigators.ip.hes_datasets import (
+    nhp_apc,
+    primary_diagnosis,
+    procedures,
+)
 
 
 def _virtual_wards_admissions(*args):
